requirements.txt
streamlit
chromadb
langchain
langchain-community
langchain-ollama
langchain-chroma
langchain-huggingface
pypdf
sentence-transformers
python-dotenv
